% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ais.R
\name{ais_enumerations}
\alias{ais_enumerations}
\title{Enums from ais.py}
\usage{
ais_enumerations()
}
\value{
tibble with ship type ids and descriptions
}
\description{
Runs python code from ais.py to return ship type legends
}
\examples{
\dontrun{
ais_enumerations()
}
}
