% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpol_ais.R
\name{interpol_ais}
\alias{interpol_ais}
\title{AIS interpolation}
\usage{
interpol_ais(x)
}
\arguments{
\item{x}{A tibble (*read_ais* output), with timestamp as first column.}
}
\value{
List with two tibbles. 1) All unique mmsis 2) complete input tibble with interpolated data
}
\description{
Interpolates missing values in tibble created by *read_ais* and selects all unique mmsis.
}
\details{
This function takes a tibble as input (create it with read_ais and add timestamps as first column)
    and outputs a list with two tibbles, one with unique info for every mmsi and one complete 
    tibble with interpolated data.
}
\section{Warning}{

Not tested for more than the datasets suggested or included in this package.
}

