% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ais.R
\name{read_ais}
\alias{read_ais}
\title{Decode ais messages}
\usage{
read_ais(file, ...)
}
\arguments{
\item{file}{a path to a file, connection or literal data}

\item{...}{params sent on to read_lines}
}
\value{
a tibble with AIS data
}
\description{
`read_ais` wraps `readr::read_lines()` and reads AIS messages 
making use of the ais.py code from \url{https://gist.github.com/ieb/9c337d68a4492db1571e} 
which has been bundled into the package. This function therefore 
has a dependency on python as stated in the SystemRequirements entry 
in the DESCRIPTION file.
}
\details{
Use the same parameters as for `readr::read_lines()`
}
