% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ais.R
\name{read_ais_log}
\alias{read_ais_log}
\title{Read a log file with lines containing comma separated Unix timestamps and AIS messages}
\usage{
read_ais_log(file, ...)
}
\arguments{
\item{file}{a path to a file, connection or literal data}

\item{...}{params sent on to readr::read_lines}
}
\value{
a tibble with the timestamp and the encoded AIS message
}
\description{
Takes a log file and parses valid rows into a tibble of timestamps and AIS messages
}
