% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scale_colour_viking}
\alias{scale_colour_viking}
\title{colour scale constructor for viking colours}
\usage{
scale_colour_viking(palette = "Alle", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in viking_palettes}

\item{discrete}{Boolean indicating whether colour aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_colour_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
colour scale constructor for viking colours
}
