% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topNorName1880to2020.R
\docType{data}
\name{topNorName1880to2020}
\alias{topNorName1880to2020}
\title{topNorName1880to2020}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{n}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
topNorName1880to2020
}
\description{
The twelve most popular Norwegian boys names between 1880 and 2020
}
\details{
DETAILS
}
\keyword{datasets}
