% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{viking_cols}
\alias{viking_cols}
\title{Function to extract viking colours as hex codes}
\usage{
viking_cols(...)
}
\arguments{
\item{...}{Character names of viking_colours}
}
\description{
Function to extract viking colours as hex codes
}
