% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{viking_pal}
\alias{viking_pal}
\title{Return function to interpolate a viking colour palette}
\usage{
viking_pal(palette = "Alle", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in viking_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a viking colour palette
}
