% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ais.R
\name{vikingr_example}
\alias{vikingr_example}
\title{Get path to readr_ais example data}
\usage{
vikingr_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If `NULL`, the example files will be listed.}
}
\description{
vikingr comes bundled with sample files in its `inst/extdata`
directory. This function make them easy to access
}
\examples{
vikingr_example()
vikingr_example("vikingr-visby-2019-ais")
vikingr_example("nmea-sample")
}
