% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ais.R
\name{read_ais}
\alias{read_ais}
\title{Decode ais messages}
\usage{
read_ais(file, ...)
}
\arguments{
\item{file}{a path to a file, connection or literal data}

\item{...}{params sent on to read_lines}
}
\value{
a tibble with AIS data
}
\description{
`read_ais` wraps `readr::read_lines()` but reads AIS messages
Use the same parameters as for `readr::read_lines()`
}
